rework createAdjacencyMatrix() and change Djikstra parameters
